+++ amazon-chime-sdk/src/main/java/com/amazonaws/services/chime/sdk/meetings/analytics/EventAttributeName.kt
@@ -104,6 +104,11 @@ enum class EventAttributeName {
     meetingErrorMessage,
 
     /**
+     * The application state
+     */
+    appState,
+
+    /**
      * Meeting Status [MeetingSessionStatus]
      */
     meetingStatus,