+++ amazon-chime-sdk/src/main/java/com/amazonaws/services/chime/sdk/meetings/ingestion/AppLifecycleObserver.kt
@@ -1,6 +1,0 @@
-package com.amazonaws.services.chime.sdk.meetings.ingestion
-
-interface AppLifecycleObserver {
-    fun startObserving()
-    fun stopObserving()
-}