+++ amazon-chime-sdk/src/test/java/com/amazonaws/services/chime/sdk/meetings/audiovideo/DefaultAudioVideoFacadeTest.kt
@@ -25,7 +25,7 @@ import com.amazonaws.services.chime.sdk.meetings.device.DeviceChangeObserver
 import com.amazonaws.services.chime.sdk.meetings.device.DeviceController
 import com.amazonaws.services.chime.sdk.meetings.device.MediaDevice
 import com.amazonaws.services.chime.sdk.meetings.device.MediaDeviceType
-import com.amazonaws.services.chime.sdk.meetings.ingestion.AppLifecycleObserver
+import com.amazonaws.services.chime.sdk.meetings.ingestion.AppStateMonitor
 import com.amazonaws.services.chime.sdk.meetings.realtime.RealtimeControllerFacade
 import com.amazonaws.services.chime.sdk.meetings.realtime.RealtimeObserver
 import com.amazonaws.services.chime.sdk.meetings.realtime.TranscriptEventObserver
@@ -93,7 +93,7 @@ class DefaultAudioVideoFacadeTest {
     private lateinit var eventAnalyticsController: EventAnalyticsController
 
     @MockK
-    private lateinit var mockAppLifecycleObserver: AppLifecycleObserver
+    private lateinit var mockAppStateMonitor: AppStateMonitor
 
     @MockK
     private lateinit var realtimeController: RealtimeControllerFacade
@@ -200,7 +200,7 @@ class DefaultAudioVideoFacadeTest {
         every { ContextCompat.checkSelfPermission(any(), any()) } returns 0
         audioVideoFacade.start()
         verify { audioVideoController.start(AudioVideoConfiguration()) }
-        verify { mockAppLifecycleObserver.startObserving() }
+        verify { mockAppStateMonitor.start() }
     }
 
     @Test
@@ -216,7 +216,7 @@ class DefaultAudioVideoFacadeTest {
     fun `stop should call audioVideoController stop`() {
         audioVideoFacade.stop()
         verify { audioVideoController.stop() }
-        verify { mockAppLifecycleObserver.stopObserving() }
+        verify { mockAppStateMonitor.stop() }
     }
 
     @Test